app/requirements.txt
SQLAlchemy>=2.0.28
psycopg2-binary>=2.9.9
pillow>=10.2.0
ffprobe_python>=1.0.3
scrapy>=2.11.1
textual-dev>=0.58.0